Fix Duplicate Block Unit problem
The type that was being used to locate node location_t was simply a unsigned integer which 
is not unique enough for node location information.  Instead use the expanded_location structure
which we can hash and perform equality operations on.   Block units should only appear once 
and many action elements can now be contained in the block unit.