Adjust C++ containment algorithm
The owner of methodunits and callable units is no longer always the
class.  The owner is the file unless the class and the function/method
are in the same file in which case we stuff the function/method into
the class.  In the case where the file of the decl and the class are
not the same we stuff a fake decl into the class to keep up
appearances